refactor: 모든 페이지에 performance tuning
